Color coding for Pymol
This script color codes CA atoms in a PDB on pymol from red to white to blue for values written in the B-factor column ([62:66]).
Only CA atoms are coloured. The script can be modified to color all atoms of a residue according to its CA's value, or by their own values if those are provided.